requirements.txt
llama-cpp-python
tk
